requirements.txt
PySide6==6.9.3
PySide6_Addons==6.9.3
PySide6_Essentials==6.9.3
python-vlc==3.0.21203
shiboken6==6.9.3
